frontend/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
bcrypt
python-jose
cryptography
yfinance
prophet
vaderSentiment
requests
pandas
numpy
pydantic